backend/requirements.txt
flask
requests
python-dotenv
gunicorn
python-dotenv
groq